Move validaor to a separate package
